test(quarto): move reopen regression test to its own file
Extract the contribution-level "cached output restore on reopen" test into
quartoOutputManagerReopen.vitest.ts with its own container, instead of
bolting a heavy workbench+contribution service stack onto the simulation-
style quartoOutputManager.vitest.ts (which only needs the base container).

- Restores the light createTestContainer().build() for the existing tests.
- Drops unused model-stub members (primaryLanguage, onDidChangeCells,
  onDidChangeLanguage) -- never accessed by the exercised path.
- beforeEach resets the mutable model state for isolation.
- Adds edge cases: already-parsed-at-reopen (no defer), stale-hash (output
  not restored), and untitled reopen (deferral generalized to any scheme).
